Update ⁨اوامر الكروب.py⁩

+++ JoKeRUB/plugins/اوامر الكروب.py
@@ -832,15 +832,15 @@ async def handle_new_message(event):
 delete_enabled = False
 aljoker_Menu = set()
 afk_start_time = datetime.now()
-
+allowed_users = set()
 @l313l.on(events.NewMessage)
 async def handle_messages(event):
-    global delete_enabled, afk_start_time
+    global delete_enabled, afk_start_time, allowed_users
     
     sender_id = event.sender_id
     current_user_id = await l313l.get_me()
     
-    if event.is_private and delete_enabled and sender_id != current_user_id.id:
+    if event.is_private and delete_enabled and sender_id != current_user_id.id and sender_id not in allowed_users:
         await event.delete()
         if sender_id not in aljoker_Menu:
             aljoker_time = aljoker_waqt()
@@ -879,19 +879,17 @@ def aljoker_waqt():
     return "N/A"
 @l313l.ar_cmd(pattern="سم")
 async def allow_user(event: Message):
-    global aljoker_Menu
+    global allowed_users
     sender_id = event.sender_id
-    if sender_id in aljoker_Menu:
-        aljoker_Menu.remove(sender_id)
-        await event.edit('**᯽︙ تم السماح للشخص بالتحدث معك في الخاص**')
+    allowed_users.add(sender_id)
+    await event.edit('**᯽︙ تم السماح للشخص بالتحدث معك في الخاص**')
 
 @l313l.ar_cmd(pattern="رف")
 async def block_user(event: Message):
-    global aljoker_Menu
+    global allowed_users
     sender_id = event.sender_id
-    if sender_id not in aljoker_Menu:
-        aljoker_Menu.add(sender_id)
-        await event.edit('**᯽︙ تم رفض الشخص من التحدث معك في الخاص**')
+    allowed_users.remove(sender_id)
+    await event.edit('**᯽︙ تم رفض الشخص من التحدث معك في الخاص**')
 points = {}
 is_game_started = False
 is_word_sent = False
